Fix error pointed out by Dave. See http://lists.admb-project.org/pipermail/users/2009-August/thread.html

+++ admb/trunk/src/df1b2-separable/df1b2fn2.cpp
@@ -88,7 +88,7 @@ int adptr_diff(void * x, void * y) { return int(x)-int(y); }
     if (sanity_flag)
       df1b2variable::pool->sanity_check();
 #endif
-    double nvar=((twointsandptr*)ptr)->ptr->nvar;
+    int nvar=((twointsandptr*)ptr)->ptr->nvar;
     int blocksize=get_blocksize(nvar);
     for (double * tmp=ptr+2;tmp<ptr+blocksize-1;*tmp++=0);
    /*
